Refactor common xcvrd functions
