+++ sonic-xcvrd/xcvrd/xcvrd_utilities/optics_si_parser.py
@@ -3,6 +3,7 @@ import os
 
 from sonic_py_common import device_info, syslogger
 from xcvrd import xcvrd
+from . import xcvrd_common
 
 g_optics_si_dict = {}
 
@@ -31,7 +32,7 @@ def get_optics_si_settings_value(physical_port, lane_speed, key, vendor_name_str
                 port_list = keys.split(COMMA_SEPARATOR)
                 for port in port_list:
                     if RANGE_SEPARATOR in port:
-                        if xcvrd.check_port_in_range(port, physical_port):
+                        if xcvrd_common.check_port_in_range(port, physical_port):
                             optics_si_dict = g_optics_si_dict[GLOBAL_MEDIA_SETTINGS_KEY][keys]
                             break
                     elif str(physical_port) == port:
@@ -39,7 +40,7 @@ def get_optics_si_settings_value(physical_port, lane_speed, key, vendor_name_str
                         break
 
             elif RANGE_SEPARATOR in keys:
-                if xcvrd.check_port_in_range(keys, physical_port):
+                if xcvrd_common.check_port_in_range(keys, physical_port):
                     optics_si_dict = g_optics_si_dict[GLOBAL_MEDIA_SETTINGS_KEY][keys]
 
             key_dict = {}
@@ -108,7 +109,7 @@ def fetch_optics_si_setting(physical_port, lane_speed, sfp):
 
     optics_si = {}
 
-    if not xcvrd._wrapper_get_presence(physical_port):
+    if not xcvrd_common._wrapper_get_presence(physical_port):
         helper_logger.log_info("Module {} presence not detected during notify".format(physical_port))
         return optics_si
     vendor_key, vendor_name = get_module_vendor_key(physical_port, sfp)
